# openapi/memphant.v1.json
{
  "components": {
    "schemas": {
      "CorrectRequest": {
        "$defs": {
          "ActorId": {
            "format": "uuid",
            "type": "string"
          },
          "CorrectSelector": {
            "properties": {
              "memory_unit_id": {
                "$ref": "#/$defs/UnitId"
              }
            },
            "required": [
              "memory_unit_id"
            ],
            "type": "object"
          },
          "CorrectionPayload": {
            "properties": {
              "reason": {
                "type": "string"
              },
              "valid_from": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "valid_to": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "value": {
                "type": "string"
              }
            },
            "required": [
              "value",
              "reason"
            ],
            "type": "object"
          },
          "ScopeId": {
            "format": "uuid",
            "type": "string"
          },
          "TenantId": {
            "format": "uuid",
            "type": "string"
          },
          "UnitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "actor_id": {
            "$ref": "#/$defs/ActorId"
          },
          "correction": {
            "$ref": "#/$defs/CorrectionPayload"
          },
          "scope_id": {
            "$ref": "#/$defs/ScopeId"
          },
          "selector": {
            "$ref": "#/$defs/CorrectSelector"
          },
          "tenant_id": {
            "$ref": "#/$defs/TenantId"
          }
        },
        "required": [
          "tenant_id",
          "scope_id",
          "actor_id",
          "selector",
          "correction"
        ],
        "title": "CorrectRequest",
        "type": "object"
      },
      "CorrectResult": {
        "$defs": {
          "UnitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "correction_id": {
            "type": "string"
          },
          "correction_kind": {
            "type": "string"
          },
          "created": {
            "items": {
              "$ref": "#/$defs/UnitId"
            },
            "type": "array"
          },
          "superseded": {
            "items": {
              "$ref": "#/$defs/UnitId"
            },
            "type": "array"
          },
          "trace_ref": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "correction_id",
          "superseded",
          "created",
          "correction_kind"
        ],
        "title": "CorrectResult",
        "type": "object"
      },
      "ErrorEnvelope": {
        "$defs": {
          "ErrorBody": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": true,
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id",
              "details"
            ],
            "type": "object"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "error": {
            "$ref": "#/$defs/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ForgetRequest": {
        "$defs": {
          "ActorId": {
            "format": "uuid",
            "type": "string"
          },
          "ForgetSelector": {
            "properties": {
              "memory_unit_id": {
                "anyOf": [
                  {
                    "$ref": "#/$defs/UnitId"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "scope_id": {
                "anyOf": [
                  {
                    "$ref": "#/$defs/ScopeId"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "type": "object"
          },
          "ScopeId": {
            "format": "uuid",
            "type": "string"
          },
          "TenantId": {
            "format": "uuid",
            "type": "string"
          },
          "UnitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "actor_id": {
            "$ref": "#/$defs/ActorId"
          },
          "reason": {
            "type": "string"
          },
          "scope_id": {
            "$ref": "#/$defs/ScopeId"
          },
          "selector": {
            "$ref": "#/$defs/ForgetSelector"
          },
          "tenant_id": {
            "$ref": "#/$defs/TenantId"
          }
        },
        "required": [
          "tenant_id",
          "scope_id",
          "actor_id",
          "selector",
          "reason"
        ],
        "title": "ForgetRequest",
        "type": "object"
      },
      "ForgetResult": {
        "$defs": {
          "UnitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "deletion_generation": {
            "format": "uint64",
            "minimum": 0,
            "type": "integer"
          },
          "invalidated_units": {
            "items": {
              "$ref": "#/$defs/UnitId"
            },
            "type": "array"
          },
          "policy": {
            "type": "string"
          },
          "trace_ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "verification": {
            "type": "string"
          }
        },
        "required": [
          "deletion_generation",
          "policy",
          "invalidated_units",
          "verification"
        ],
        "title": "ForgetResult",
        "type": "object"
      },
      "MarkRequest": {
        "$defs": {
          "MarkOutcome": {
            "enum": [
              "success",
              "failure",
              "corrected",
              "ignored"
            ],
            "type": "string"
          },
          "TenantId": {
            "format": "uuid",
            "type": "string"
          },
          "TraceId": {
            "format": "uuid",
            "type": "string"
          },
          "UnitId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "caller_id": {
            "type": "string"
          },
          "outcome": {
            "$ref": "#/$defs/MarkOutcome"
          },
          "tenant_id": {
            "$ref": "#/$defs/TenantId"
          },
          "trace_id": {
            "$ref": "#/$defs/TraceId"
          },
          "used_ids": {
            "items": {
              "$ref": "#/$defs/UnitId"
            },
            "type": "array"
          }
        },
        "required": [
          "tenant_id",
          "trace_id",
          "caller_id",
          "used_ids",
          "outcome"
        ],
        "title": "MarkRequest",
        "type": "object"
      },
      "MarkResult": {
        "$defs": {
          "TraceId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "trace_id": {
            "$ref": "#/$defs/TraceId"
          }
        },
        "required": [
          "accepted",
          "trace_id"
        ],
        "title": "MarkResult",
        "type": "object"
      },
      "RecallHttpRequest": {
        "$defs": {
          "ActorId": {
            "format": "uuid",
            "type": "string"
          },
          "RecallMode": {
            "enum": [
              "fast",
              "balanced",
              "exhaustive"
            ],
            "type": "string"
          },
          "ScopeId": {
            "format": "uuid",
            "type": "string"
          },
          "TenantId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "actor_id": {
            "$ref": "#/$defs/ActorId"
          },
          "allowed_scope_ids": {
            "items": {
              "$ref": "#/$defs/ScopeId"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "budget_tokens": {
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "include_beliefs": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "include_trace": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "limit": {
            "format": "uint",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "mode": {
            "anyOf": [
              {
                "$ref": "#/$defs/RecallMode"
              },
              {
                "type": "null"
              }
            ]
          },
          "query": {
            "type": "string"
          },
          "scope_id": {
            "$ref": "#/$defs/ScopeId"
          },
          "tenant_id": {
            "$ref": "#/$defs/TenantId"
          }
        },
        "required": [
          "tenant_id",
          "scope_id",
          "actor_id",
          "query"
        ],
        "title": "RecallHttpRequest",
        "type": "object"
      },
...
      "RetainEpisodeHttpRequest": {
        "$defs": {
          "ActorId": {
            "format": "uuid",
            "type": "string"
          },
          "ScopeId": {
            "format": "uuid",
            "type": "string"
          },
          "TenantId": {
            "format": "uuid",
            "type": "string"
          },
          "TrustLevel": {
            "enum": [
              "trusted_user",
              "trusted_system",
              "verified_tool",
              "unverified_tool",
              "web_content",
              "agent_output",
              "imported_external",
              "quarantined"
            ],
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "actor_id": {
            "$ref": "#/$defs/ActorId"
          },
          "body": {
            "type": "string"
          },
          "compiler_version": {
            "type": [
              "string",
              "null"
            ]
          },
          "scope_id": {
            "$ref": "#/$defs/ScopeId"
          },
          "source_kind": {
            "type": "string"
          },
          "source_trust": {
            "$ref": "#/$defs/TrustLevel"
          },
          "subject_hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "tenant_id": {
            "$ref": "#/$defs/TenantId"
          }
        },
        "required": [
          "tenant_id",
          "scope_id",
          "actor_id",
          "source_kind",
          "source_trust",
          "body"
        ],
        "title": "RetainEpisodeHttpRequest",
        "type": "object"
      },
      "RetainEpisodeHttpResponse": {
        "$defs": {
          "DedupOutcome": {
            "properties": {
              "matched": {
                "type": "boolean"
              },
              "observation_count": {
                "format": "uint32",
                "minimum": 0,
                "type": "integer"
              }
            },
            "required": [
              "matched",
              "observation_count"
            ],
            "type": "object"
          },
          "EpisodeId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "properties": {
          "dedup": {
            "$ref": "#/$defs/DedupOutcome"
          },
          "enqueued": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "episode_id": {
            "$ref": "#/$defs/EpisodeId"
          },
          "trace_ref": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "episode_id",
          "dedup",
          "enqueued"
        ],
        "title": "RetainEpisodeHttpResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MemPhant API",
    "version": "0.1.0-ws0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/correct": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectResult"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/episodes": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetainEpisodeHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetainEpisodeHttpResponse"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/forget": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetResult"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/mark": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarkResult"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/memory": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetainEpisodeHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetainEpisodeHttpResponse"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recall": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallHttpRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallResponse"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reflect": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReflectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReflectResult"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scopes/{id}/block": {
      "get": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scopes/{id}/memory": {
      "get": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScopeMemoryResponse"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/scopes/{id}/stats": {
      "get": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopeId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/traces/{id}": {
      "get": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraceId"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalTrace"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    }
  }
}